web_jalivur/requirements.txt
annotated-types==0.6.0
bcrypt==4.1.2
cryptography==41.0.7
cx-Freeze==6.15.10
Django==4.2.6
easygui==0.98.3
email-validator==2.1.0.post1
flask==3.0.0
gmpy2==2.1.5
google==3.0.0
google-auth==2.25.2
httptools==0.6.1
Mypackage==0.1
mysql-connector-python==8.2.0
orjson==3.9.10
pafy==0.5.5
pandas==2.1.3
passlib==1.7.4
pip==23.3.1
psycopg2-binary==2.9.9
py2app==0.28.6
pydantic-core==2.10.1
pydantic-extra-types==2.1.0
pydantic-settings==2.1.0
PyInquirer==1.0.3
pyinstaller==6.1.0
pymongo==4.6.1
PyMySQL==1.1.0
python-jose==3.3.0
pytube==15.0.0
PyYAML==6.0.1
qrcode==7.4.2
qrcodegen==1.8.0
reflex==0.3.7
reportlab==4.0.7
tkinterdnd2==0.3.0
ujson==5.8.0
uvloop==0.19.0
virtualenv==20.24.5
youtube-dl==2020.12.2
